Fixed precision issues

+++ ldm/models/diffusion/plms.py
@@ -18,7 +18,7 @@ class PLMSSampler(object):
     def register_buffer(self, name, attr):
         if type(attr) == torch.Tensor:
             if attr.device != torch.device("cpu"):
-                attr = attr.to(torch.device("cpu"))
+                attr = attr.to(torch.float32).to(torch.device("cpu"))
         setattr(self, name, attr)
 
     def make_schedule(self, ddim_num_steps, ddim_discretize="uniform", ddim_eta=0., verbose=True):
